Moved DisplacedVertexCluster files into plugins/.
