--- RecoTracker/DisplacedRegionalTracking/BuildFile.xml
@@ -1,11 +1,0 @@
-<use name="DataFormats/Candidate"/>
-<use name="DataFormats/Common"/>
-<use name="DataFormats/GeometrySurface"/>
-<use name="DataFormats/GeometryVector"/>
-<use name="DataFormats/Math"/>
-<use name="TrackingTools/GeomPropagators"/>
-<use name="TrackingTools/TrajectoryState"/>
-<use name="TrackingTools/TransientTrack"/>
-<export>
-  <lib name="1"/>
-</export>
